Hoist shared worker-client state into client/core, client/handles, client/snapshots
Lift the transport singleton (RPC correlation, pending/subscriber maps, port
wiring, the Studio auth-lens default and op-issuer stamp), the handle types,
and the firestore snapshot machinery out of client.ts into three sibling
modules under client/. client.ts imports the machinery for its API-family
functions and re-exports the shared public types + lens/issuer controls, so the
serve/worker/client export surface is byte-identical.

_defaultLens is exported as a live binding so listener families read the same
singleton lens value they did when it was inline. No behavior change; pure move.
